Create a dataset for simulation of multiple dose of caffeine

DatasetMulti will create a dataset for simulation of multiple dose of caffeine

DatasetMulti(Weight, Dose, N = 20, Tau = 24)

Arguments

Weight
Body weight (kg)

Dose
Dose of multiple caffeine (mg)

N
The number of simulated subjects

Tau
The interval of multiple dosing (hour)

Value

The dataset of pharmacokinetic parameters of subjects after multiple caffeine dose following multivariate normal

See also

http://asancpt.github.io/CaffeineEdison

Examples

DatasetMulti(Weight = 20, Dose = 200, N = 20, Tau = 8)

#>        TmaxS     CmaxS      AUCS       AI    Aavss     Cavss   Cmaxss    Cminss
#> 1  1.0465498  8.921906  69.88118 1.434696 167.1472  8.735147 14.96431  4.534010
#> 2  0.2041821 11.496640  89.41407 1.532771 188.8675 11.176759 18.10346  6.292525
#> 3  3.2675552 10.413368 166.35163 2.082768 305.0916 20.793954 28.33175 14.728813
#> 4  0.9152337 13.048549  96.88571 1.406125 160.7044 12.110714 21.14909  6.108395
#> 5  1.7980280 14.220256 163.57209 1.759665 237.5996 20.446512 30.22242 13.047316
#> 6  0.3247730 13.009918  78.09987 1.322480 141.4264  9.762483 18.21985  4.442822
#> 7  0.8668605  9.423756  64.26030 1.345566 146.8193  8.032538 14.69267  3.773348
#> 8  1.0942722 12.216377 109.71266 1.559707 194.7466 13.714082 21.92127  7.866537
#> 9  0.6185416 12.577403  97.60394 1.481048 177.4809 12.200493 20.31986  6.599942
#> 10 0.4210418 13.725909 121.78979 1.632978 210.5927 15.223724 23.56045  9.132544
#> 11 1.1602853  7.883181  73.94273 1.597266 202.8940  9.242841 14.52243  5.430375
#> 12 1.5462671  8.835975 108.93661 1.895030 266.0649 13.617076 19.35700  9.142385
#> 13 2.0167570  8.743906  74.11363 1.367137 151.8051  9.264203 16.65174  4.471734
#> 14 0.9507021  9.805622  53.99031 1.198189 110.9203  6.748789 14.55009  2.406689
#> 15 1.4676649 11.262517  75.48271 1.252578 124.6434  9.435339 18.92425  3.816004
#> 16 2.1146148  7.435544  89.63089 1.779438 241.7793 11.203861 16.45724  7.208678
#> 17 1.4716263 15.800664 204.32401 1.978225 283.4132 25.540501 35.58039 17.594372
#> 18 1.3996019 14.131917 141.07121 1.635202 211.0708 17.633901 27.26575 10.591509
#> 19 0.2897304 10.876492  70.25021 1.375545 153.7359  8.781276 15.68133  4.281248
#> 20 2.7810179  8.099173  85.46025 1.485396 178.4434 10.682531 17.74767  5.799562

DatasetMulti(20,500)

#>        TmaxS    CmaxS     AUCS       AI     Aavss     Cavss   Cmaxss     Cminss
#> 1  0.5506126 29.46363 258.6709 1.056808 170.68243 10.777953 33.29734  1.7898890
#> 2  2.1992528 43.63522 879.3395 1.351167 370.29614 36.639146 66.70692 17.3370654
#> 3  1.5998776 40.59759 569.5358 1.166263 256.14127 23.730657 53.91288  7.6858573
#> 4  0.3986528 33.79385 196.7727 1.011973 112.45363  8.198862 36.81413  0.4355495
#> 5  0.4896296 29.19146 154.5894 1.006643  99.37872  6.441225 32.55490  0.2148392
#> 6  1.2757148 39.18317 420.7965 1.083722 194.85642 17.533188 48.65526  3.7588206
#> 7  0.8021400 51.48487 501.4773 1.072329 185.04950 20.894888 60.41517  4.0750401
#> 8  0.5056932 41.46803 385.9691 1.069674 182.68389 16.082046 46.98491  3.0604171
#> 9  0.5665050 40.28008 276.7934 1.022366 130.53773 11.533058 45.06926  0.9859569
#> 10 1.6274096 20.05760 156.6567 1.017881 123.45260  6.527361 26.85347  0.4717443
#> 11 0.1926712 35.31528 185.1291 1.008671 104.90186  7.713713 37.00798  0.3181247
#> 12 2.3764303 33.71089 307.2579 1.021559 129.32091 12.802413 50.46057  1.0649026
#> 13 0.8883006 27.94080 293.0118 1.088311 198.66887 12.208824 33.37051  2.7078549
#> 14 0.9145419 26.33233 175.6727 1.014827 118.06798  7.319698 31.39198  0.4586410
#> 15 0.2244007 42.51075 293.8814 1.028376 138.97904 12.245060 45.20947  1.2474792
#> 16 1.0625193 34.24628 330.0741 1.063292 176.85054 13.753087 41.25837  2.4558884
#> 17 0.3122303 32.04957 177.6950 1.010201 108.57847  7.403960 34.37115  0.3470909
#> 18 1.4668076 45.11673 385.3804 1.031971 143.61061 16.057519 57.57383  1.7836701
#> 19 2.3087835 31.70246 352.9656 1.062702 176.30006 14.706899 44.23295  2.6098564
#> 20 0.8688179 33.70054 214.8564 1.012216 112.95873  8.952349 40.02727  0.4830536