Create a dataset for simulation of single dose of caffeine

caffDataset will create a dataset for simulation of single dose of caffeine

caffDataset(Weight, Dose, N = 20)

Arguments

Weight
Body weight (kg)

Dose
Dose of single caffeine (mg)

N
The number of simulated subjects

Value

The dataset of pharmacokinetic parameters of subjects after single caffeine dose following multivariate normal

See also

https://asancpt.github.io/caffsim

Examples

caffDataset(Weight = 20, Dose = 200, N = 20)

#>         Tmax      Cmax       AUC Half_life        CL         V        Ka
#> 1  1.5667033  7.707605  66.80369  4.787710 2.9938466 20.683504  1.727243
#> 2  0.3601398 13.305581 158.49765  8.001588 1.2618484 14.569684 14.258002
#> 3  0.8510911 11.781636 105.22666  5.567288 1.9006590 15.269142  4.281389
#> 4  0.7782195 15.438372 139.20991  5.683152 1.4366794 11.781914  4.856538
#> 5  2.2774760 10.471597 153.37612  8.414262 1.3039840 15.832702  1.290616
#> 6  1.4845261 11.615494  98.84250  4.748383 2.0234211 13.864328  1.860576
#> 7  1.1966568  8.817884  61.42177  3.903190 3.2561744 18.339782  2.328228
#> 8  2.3583556 10.905700 117.71047  5.581072 1.6990841 13.683565  1.015073
#> 9  2.0434872 12.873511 251.97532 12.061580 0.7937285 13.814748  1.721153
#> 10 1.7465542 11.186493 116.65817  5.883074 1.7144105 14.554118  1.617829
#> 11 0.8975993 18.276502 236.38241  8.317151 0.8460866 10.154445  4.536546
#> 12 0.8214509  9.511383 100.71694  6.744262 1.9857632 19.325408  4.776292
#> 13 0.1457276 10.457600  62.98155  4.071383 3.1755333 18.656297 37.122197
#> 14 0.9476406 11.421522  71.03058  3.589142 2.8156886 14.582836  3.134025
#> 15 0.5363007 11.440509  85.35123  4.783610 2.3432585 16.174941  7.505612
#> 16 0.3612881 14.622818 131.99674  5.999878 1.5151890 13.118252 13.239938
#> 17 0.2753061 14.233422 103.42048  4.840762 1.9338529 13.508402 17.626186
#> 18 1.0709614 13.573473 154.36926  7.099034 1.2955947 13.271964  3.417666
#> 19 0.2391135 20.594838 157.53699  5.132586 1.2695431  9.402653 21.301091
#> 20 0.4366808  8.520508  73.66449  5.680541 2.7150124 22.255036 10.274596
#>            Ke
#> 1  0.14474562
#> 2  0.08660781
#> 3  0.12447713
#> 4  0.12193938
#> 5  0.08236017
#> 6  0.14594441
#> 7  0.17754706
#> 8  0.12416970
#> 9  0.05745516
#> 10 0.11779556
#> 11 0.08332180
#> 12 0.10275401
#> 13 0.17021241
#> 14 0.19308237
#> 15 0.14486968
#> 16 0.11550236
#> 17 0.14315927
#> 18 0.09761891
#> 19 0.13501967
#> 20 0.12199542

caffDataset(20,500)

#>         Tmax     Cmax      AUC Half_life       CL         V        Ka
#> 1  1.2194891 31.23224 399.3750  7.970037 1.251956 14.398467  2.987119
#> 2  0.3213105 33.32380 210.9643  4.158464 2.370070 14.222005 13.944452
#> 3  0.4962064 17.44264 111.8279  4.084188 4.471154 26.350700  7.913188
#> 4  2.6033963 33.20029 366.8756  5.524279 1.362860 10.864094  0.868794
#> 5  0.6024500 40.59588 490.1593  7.938690 1.020077 11.685529  7.473537
#> 6  0.7435651 26.14265 160.8689  3.711604 3.108121 16.646630  4.452045
#> 7  1.8569474 24.77113 261.6656  5.881897 1.910836 16.218385  1.480972
#> 8  0.5884100 24.21825 139.2015  3.551118 3.591916 18.405942  6.023966
#> 9  0.1622126 38.96229 368.9296  6.448540 1.355272 12.611150 35.939555
#> 10 0.2775437 27.88873 194.1060  4.626899 2.575913 17.198396 17.251659
#> 11 0.4162273 37.21340 497.4143  8.969877 1.005198 13.010829 12.248494
#> 12 0.9812631 24.31673 231.4308  5.874586 2.160473 18.314405  3.602136
#> 13 2.1074704 33.81056 441.7699  7.441103 1.131811 12.152846  1.368252
#> 14 0.4640463 24.57787 257.5687  6.933275 1.941230 19.421469 10.031748
#> 15 0.9444475 45.91698 443.0721  5.995482 1.128485  9.763072  3.819060
#> 16 0.2123997 35.38782 223.4953  4.226925 2.237183 13.645607 23.550802
#> 17 0.8490611 28.92339 190.3248  3.925365 2.627088 14.880635  3.787447
#> 18 0.9678986 21.90591 211.0649  5.967194 2.368940 20.398160  3.689267
#> 19 0.2426456 26.96820 123.1923  2.992697 4.058695 17.527338 18.223227
#> 20 0.1936110 40.74015 265.2176  4.375165 1.885245 11.902246 26.626877
#>            Ke
#> 1  0.08695066
#> 2  0.16664806
#> 3  0.16967877
#> 4  0.12544624
#> 5  0.08729400
#> 6  0.18671170
#> 7  0.11781914
#> 8  0.19514980
#> 9  0.10746619
#> 10 0.14977634
#> 11 0.07725858
#> 12 0.11796577
#> 13 0.09313135
#> 14 0.09995277
#> 15 0.11558704
#> 16 0.16394898
#> 17 0.17654408
#> 18 0.11613499
#> 19 0.23156370
#> 20 0.15839402