Create a dataset of pharmacokinetic parameters of single oral dosing of caffeine

caffPkparam will create a dataset for simulation of single dose of caffeine

caffPkparam(Weight, Dose, N = 20)

Arguments

Weight
Body weight (kg)

Dose
Dose of single caffeine (mg)

N
The number of simulated subjects

Value

The dataset of pharmacokinetic parameters of subjects after single caffeine dose following multivariate normal

See also

Examples

caffPkparam(Weight = 20, Dose = 200, N = 20)

#> # A tibble: 20 x 9
#>    subjid  Tmax  Cmax   AUC Half_life    CL     V     Ka     Ke
#>     <int> <dbl> <dbl> <dbl>     <dbl> <dbl> <dbl>  <dbl>  <dbl>
#>  1      1 0.682 10.6   78.2      4.61  2.56 17.0   5.41  0.150 
#>  2      2 0.169 16.8  103.       4.12  1.95 11.6  30.9   0.168 
#>  3      3 1.57  11.5   93.0      4.37  2.15 13.6   1.65  0.158 
#>  4      4 0.681 21.9  160.       4.58  1.25  8.23  5.40  0.151 
#>  5      5 2.44  12.9  139.       5.53  1.43 11.5   0.960 0.125 
#>  6      6 0.810 10.9   61.5      3.30  3.25 15.5   3.78  0.210 
#>  7      7 0.555 13.6  166.       8.08  1.20 14.0   8.32  0.0858
#>  8      8 2.22   7.84  76.2      4.93  2.62 18.7   1.04  0.140 
#>  9      9 1.71  14.1  170.       7.05  1.18 12.0   1.80  0.0983
#> 10     10 0.469 12.1  115.       6.21  1.74 15.6   9.61  0.112 
#> 11     11 0.982 16.9  160.       5.81  1.25 10.5   3.58  0.119 
#> 12     12 0.606 12.7   70.7      3.40  2.83 13.9   5.70  0.204 
#> 13     13 2.10  15.5  199.       7.29  1.00 10.6   1.37  0.0951
#> 14     14 0.412 11.6   69.3      3.86  2.89 16.1   9.91  0.180 
#> 15     15 0.556 12.7  130.       6.67  1.54 14.8   7.89  0.104 
#> 16     16 0.791 13.2   93.5      4.33  2.14 13.4   4.32  0.160 
#> 17     17 1.92  11.1   98.1      4.58  2.04 13.5   1.25  0.151 
#> 18     18 0.913 12.0   90.5      4.56  2.21 14.5   3.63  0.152 
#> 19     19 0.719 12.0   76.4      3.89  2.62 14.7   4.74  0.178 
#> 20     20 1.89   9.13  92.3      5.53  2.17 17.3   1.41  0.125 

caffPkparam(20,500)

#> # A tibble: 20 x 9
#>    subjid  Tmax  Cmax   AUC Half_life    CL     V     Ka     Ke
#>     <int> <dbl> <dbl> <dbl>     <dbl> <dbl> <dbl>  <dbl>  <dbl>
#>  1      1 1.95   21.1 193.       4.78 2.59  17.8   1.25  0.145 
#>  2      2 0.439  37.6 249.       4.28 2.01  12.4   9.42  0.162 
#>  3      3 1.11   27.7 249.       5.41 2.01  15.7   2.96  0.128 
#>  4      4 0.123  43.4 469.       7.40 1.07  11.4  51.3   0.0936
#>  5      5 1.14   34.0 340.       6.08 1.47  12.9   2.98  0.114 
#>  6      6 0.652  29.8 314.       6.83 1.59  15.7   6.48  0.101 
#>  7      7 1.04   50.1 494.       6.08 1.01   8.87  3.37  0.114 
#>  8      8 3.83   28.0 417.       7.08 1.20  12.3   0.547 0.0979
#>  9      9 0.367  16.1  99.9      4.04 5.00  29.2  11.7   0.171 
#> 10     10 1.71   30.8 276.       4.88 1.81  12.7   1.54  0.142 
#> 11     11 1.63   30.1 365.       7.18 1.37  14.2   1.94  0.0965
#> 12     12 1.07   43.0 469.       6.78 1.07  10.4   3.39  0.102 
#> 13     13 3.54   20.4 265.       5.98 1.88  16.3   0.562 0.116 
#> 14     14 0.761  27.9 180.       3.90 2.78  15.6   4.39  0.178 
#> 15     15 0.511  34.4 241.       4.49 2.07  13.4   7.83  0.154 
#> 16     16 1.41   32.5 348.       6.36 1.44  13.2   2.25  0.109 
#> 17     17 0.347  21.5 192.       5.92 2.61  22.3  13.9   0.117 
#> 18     18 0.829  46.5 547.       7.57 0.913  9.97  4.89  0.0916
#> 19     19 0.955  29.6 165.       3.12 3.03  13.7   2.92  0.222 
#> 20     20 0.953  42.0 265.       3.65 1.89   9.92  3.13  0.190