Create a dataset of pharmacokinetic parameters of multiple oral dosing of caffeine

caffPkparamMulti will create a dataset for simulation of multiple dose of caffeine.

caffPkparamMulti(Weight, Dose, N = 20, Tau = 8)

Arguments

Weight
Body weight (kg)

Dose
Dose of multiple caffeine (mg)

N
The number of simulated subjects

Tau
The interval of multiple dosing (hour)

Value

The dataset of pharmacokinetic parameters of subjects after multiple caffeine dose following multivariate normal

See also

Examples

caffPkparamMulti(Weight = 20, Dose = 200, N = 20, Tau = 8)

#> # A tibble: 20 x 9
#>    subjid TmaxS CmaxS  AUCS    AI Aavss Cavss Cmaxss Cminss
#>     <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>
#>  1      1 0.454 15.4  167.   1.86  259. 20.8    29.9  13.8 
#>  2      2 0.281 13.3   90.2  1.41  162. 11.3    19.6   5.73
#>  3      3 0.664 12.2   88.5  1.42  163. 11.1    19.2   5.64
#>  4      4 0.530 14.8  104.   1.41  162. 13.0    22.6   6.60
#>  5      5 2.89   7.45  98.2  1.79  245. 12.3    17.9   7.95
#>  6      6 0.500 10.6   65.0  1.32  141.  8.13   15.2   3.68
#>  7      7 1.96  14.9  177.   1.78  242. 22.1    32.4  14.2 
#>  8      8 1.52   9.65 118.   1.89  264. 14.7    21.0   9.87
#>  9      9 1.81  11.6  120.   1.61  206. 15.0    23.4   8.89
#> 10     10 1.66  10.3   95.2  1.51  184. 11.9    19.5   6.59
#> 11     11 0.867 12.8  145.   1.87  261. 18.1    25.9  12.1 
#> 12     12 1.46   9.99  85.1  1.46  172. 10.6    18.0   5.63
#> 13     13 0.831 15.7  183.   1.91  269. 22.9    32.4  15.4 
#> 14     14 1.48   9.03  95.1  1.69  222. 11.9    18.0   7.34
#> 15     15 0.680 16.5  181.   1.84  255. 22.6    32.6  14.9 
#> 16     16 0.561 10.2   78.5  1.48  177.  9.81   16.4   5.30
#> 17     17 0.463 17.0  102.   1.30  137. 12.7    24.2   5.64
#> 18     18 1.75   8.86  81.8  1.50  181. 10.2    16.9   5.60
#> 19     19 0.829 11.9   71.9  1.27  129.  8.99   17.7   3.76
#> 20     20 0.732 12.0  123.   1.75  236. 15.4    22.8   9.79

caffPkparamMulti(20,500)

#> # A tibble: 20 x 9
#>    subjid TmaxS CmaxS  AUCS    AI Aavss Cavss Cmaxss Cminss
#>     <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>
#>  1      1 1.27   34.2 320.   1.58  500.  40.1   63.3  23.3 
#>  2      2 1.32   25.4 228.   1.53  470.  28.5   46.3  16.0 
#>  3      3 0.624  44.4 590.   2.13  789.  73.7   99.4  52.8 
#>  4      4 0.266  22.3 131.   1.31  349.  16.3   30.7   7.35
#>  5      5 0.505  21.9 158.   1.43  417.  19.7   33.8  10.2 
#>  6      6 0.558  21.3 106.   1.19  274.  13.3   28.8   4.67
#>  7      7 0.301  28.0 271.   1.74  583.  33.8   50.3  21.4 
#>  8      8 1.22   29.2 325.   1.80  613.  40.7   59.4  26.3 
#>  9      9 0.670  37.4 341.   1.63  525.  42.6   66.0  25.5 
#> 10     10 2.53   24.7 271.   1.58  497.  33.8   53.6  19.6 
#> 11     11 0.609  19.6 132.   1.37  381.  16.5   29.6   8.00
#> 12     12 1.57   23.3 189.   1.39  394.  23.6   41.6  11.7 
#> 13     13 3.15   31.4 440.   1.85  644.  55.0   79.1  36.4 
#> 14     14 1.51   14.1  98.6  1.28  328.  12.3   24.0   5.25
#> 15     15 0.282  39.3 258.   1.39  392.  32.3   57.1  16.0 
#> 16     16 1.15   22.3 167.   1.39  390.  20.9   37.1  10.3 
#> 17     17 3.90   21.0 308.   1.81  619.  38.5   56.1  25.1 
#> 18     18 0.433  35.0 191.   1.26  313.  23.9   47.9   9.73
#> 19     19 0.932  30.3 221.   1.39  392.  27.6   48.8  13.7 
#> 20     20 1.10   19.1 106.   1.18  268.  13.3   29.3   4.55