Create a dataset for simulation of single dose of caffeine

caffPkparam will create a dataset for simulation of single dose of caffeine

caffPkparam(Weight, Dose, N = 20)

Arguments

Weight
Body weight (kg)

Dose
Dose of single caffeine (mg)

N
The number of simulated subjects

Value

The dataset of pharmacokinetic parameters of subjects after single caffeine dose following multivariate normal

See also

https://asancpt.github.io/caffsim

Examples

caffPkparam(Weight = 20, Dose = 200, N = 20)

#> # A tibble: 20 x 9
#>    subjid      Tmax      Cmax       AUC Half_life       CL         V         Ka
#>     <int>     <dbl>     <dbl>     <dbl>     <dbl>    <dbl>     <dbl>      <dbl>
#>  1      1 0.6815481 10.610739  78.20646  4.610404 2.557333 17.013477  5.4071000
#>  2      2 0.1694741 16.804692 102.75059  4.118142 1.946461 11.566814 30.9339220
#>  3      3 1.5710684 11.488799  92.99001  4.372848 2.150769 13.571406  1.6496186
#>  4      4 0.6811672 21.919667 160.46253  4.575837 1.246397  8.229884  5.3975246
#>  5      5 2.4401053 12.863231 139.39826  5.532100 1.434738 11.453268  0.9597418
#>  6      6 0.8101356 10.882741  61.51036  3.304979 3.251485 15.506620  3.7789715
#>  7      7 0.5553883 13.602520 166.28455  8.077405 1.202758 14.018990  8.3229345
#>  8      8 2.2193872  7.839475  76.24080  4.934916 2.623267 18.680526  1.0445897
#>  9      9 1.7117636 14.087598 169.53124  7.047684 1.179724 11.997575  1.7951140
#> 10     10 0.4691640 12.134696 114.66068  6.214367 1.744277 15.641527  9.6101287
#> 11     11 0.9819875 16.931253 159.54174  5.808091 1.253590 10.506447  3.5842571
#> 12     12 0.6060609 12.718825  70.66952  3.403497 2.830074 13.899204  5.7019745
#> 13     13 2.0976086 15.506734 199.04914  7.286779 1.004777 10.565062  1.3651436
#> 14     14 0.4121059 11.557513  69.27030  3.857088 2.887240 16.069755  9.9108198
#> 15     15 0.5563496 12.732747 129.86167  6.671018 1.540100 14.825451  7.8860207
#> 16     16 0.7913932 13.186071  93.45329  4.326770 2.140107 13.361832  4.3248879
#> 17     17 1.9177109 11.108858  98.10531  4.578100 2.038626 13.467578  1.2538461
#> 18     18 0.9127606 11.968976  90.48517  4.560554 2.210307 14.545777  3.6280847
#> 19     19 0.7194890 11.968891  76.44736  3.894378 2.616179 14.701863  4.7399030
#> 20     20 1.8863042  9.131570  92.31756  5.531439 2.166435 17.292212  1.4077880
#> # ... with 1 more variables: Ke <dbl>

caffPkparam(20,500)

#> # A tibble: 20 x 9
#>    subjid      Tmax     Cmax       AUC Half_life        CL         V         Ka
#>     <int>     <dbl>    <dbl>     <dbl>     <dbl>     <dbl>     <dbl>      <dbl>
#>  1      1 1.9454813 21.13020 193.21449  4.779185 2.5877977 17.846413  1.2538279
#>  2      2 0.4387019 37.56361 248.86438  4.276137 2.0091264 12.397258  9.4234136
#>  3      3 1.1086791 27.67530 249.11127  5.412317 2.0071352 15.675689  2.9611134
#>  4      4 0.1232268 43.37299 468.53392  7.400197 1.0671586 11.395648 51.2607518
#>  5      5 1.1377766 34.02069 339.66595  6.077064 1.4720345 12.908583  2.9829199
#>  6      6 0.6516363 29.76417 313.51130  6.832650 1.5948389 15.724351  6.4813008
#>  7      7 1.0397441 50.05456 494.15797  6.076561 1.0118222  8.872150  3.3709646
#>  8      8 3.8309176 28.04247 416.66666  7.075377 1.2000000 12.251735  0.5468741
#>  9      9 0.3665183 16.08659  99.93516  4.042997 5.0032443 29.189181 11.6923846
#> 10     10 1.7087332 30.78990 276.38911  4.880634 1.8090438 12.740663  1.5352345
#> 11     11 1.6257622 30.12170 365.07472  7.179286 1.3695826 14.188493  1.9432102
#> 12     12 1.0651067 42.99016 469.13322  6.782634 1.0657953 10.431313  3.3900437
#> 13     13 3.5382221 20.39996 265.25744  5.979905 1.8849613 16.265353  0.5622465
#> 14     14 0.7612997 27.93256 180.02725  3.901516 2.7773574 15.636226  4.3910093
#> 15     15 0.5114560 34.41342 241.29515  4.490276 2.0721510 13.426449  7.8321785
#> 16     16 1.4139333 32.47826 347.51777  6.355684 1.4387754 13.195384  2.2498195
#> 17     17 0.3469954 21.53754 191.61730  5.920131 2.6093677 22.291197 13.8794740
#> 18     18 0.8287644 46.47351 547.41778  7.566247 0.9133792  9.972370  4.8913845
#> 19     19 0.9549840 29.59051 164.86649  3.123991 3.0327570 13.671437  2.9211632
#> 20     20 0.9534631 42.03426 265.11656  3.646457 1.8859629  9.923641  3.1274414
#> # ... with 1 more variables: Ke <dbl>