Create a dataset for simulation of multiple dose of caffeine

caffPkparamMulti will create a dataset for simulation of multiple dose of caffeine

caffPkparamMulti(Weight, Dose, N = 20, Tau = 8)

Arguments

Weight
Body weight (kg)

Dose
Dose of multiple caffeine (mg)

N
The number of simulated subjects

Tau
The interval of multiple dosing (hour)

Value

The dataset of pharmacokinetic parameters of subjects after multiple caffeine dose following multivariate normal

See also

https://asancpt.github.io/caffsim

Examples

caffPkparamMulti(Weight = 20, Dose = 200, N = 20, Tau = 8)

#> # A tibble: 20 x 9
#>    subjid     TmaxS     CmaxS      AUCS       AI    Aavss     Cavss   Cmaxss
#>     <int>     <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>    <dbl>
#>  1      1 0.4540524 15.363198 166.66317 1.861585 259.0621 20.832896 29.87819
#>  2      2 0.2814558 13.283017  90.20407 1.412929 162.2443 11.275509 19.59804
#>  3      3 0.6639734 12.239828  88.47994 1.415872 162.9092 11.059992 19.18487
#>  4      4 0.5300914 14.772837 104.03146 1.411432 161.9059 13.003933 22.62548
#>  5      5 2.8933948  7.449264  98.20559 1.794458 244.9490 12.275699 17.94856
#>  6      6 0.4998679 10.556732  65.01033 1.318941 140.5940  8.126292 15.21518
#>  7      7 1.9645077 14.850701 176.59134 1.782099 242.3413 22.073917 32.39735
#>  8      8 1.5174001  9.645870 117.96312 1.887064 264.3986 14.745390 21.00436
#>  9      9 1.8135914 11.638279 119.85527 1.613198 206.3340 14.981909 23.37813
#> 10     10 1.6551794 10.330604  95.24749 1.509297 183.7167 11.905936 19.52159
#> 11     11 0.8669219 12.772225 145.01401 1.869426 260.7056 18.126752 25.94202
#> 12     12 1.4558914  9.990074  85.08727 1.456595 172.0464 10.635908 17.97189
#> 13     13 0.8309900 15.706750 182.87174 1.908665 268.9149 22.858967 32.38152
#> 14     14 1.4840294  9.033248  95.09902 1.687929 222.3600 11.887378 18.00982
#> 15     15 0.6801345 16.535619 180.89720 1.844049 255.3830 22.612150 32.58727
#> 16     16 0.5606221 10.221614  78.51277 1.479209 177.0733  9.814096 16.36260
#> 17     17 0.4633479 17.020010 101.84146 1.304664 137.2192 12.730182 24.15708
#> 18     18 1.7451343  8.856444  81.76839 1.497509 181.1199 10.221049 16.86649
#> 19     19 0.8290299 11.869237  71.93819 1.269464 128.7710  8.992274 17.69285
#> 20     20 0.7315648 12.042195 123.13679 1.752719 236.1294 15.392099 22.80268
#> # ... with 1 more variables: Cminss <dbl>

caffPkparamMulti(20,500)

#> # A tibble: 20 x 9
#>    subjid     TmaxS    CmaxS      AUCS       AI    Aavss    Cavss   Cmaxss
#>     <int>     <dbl>    <dbl>     <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1      1 1.2653275 34.16390 320.48919 1.583149 499.5955 40.06115 63.34211
#>  2      2 1.3153654 25.44225 228.24669 1.528617 469.8951 28.53084 46.31036
#>  3      3 0.6235091 44.36327 589.53304 2.133559 788.9576 73.69163 99.43397
#>  4      4 0.2655234 22.27586 130.68525 1.314631 348.9446 16.33566 30.70786
#>  5      5 0.5048797 21.87322 157.67440 1.432951 416.8887 19.70930 33.80245
#>  6      6 0.5575835 21.28906 106.23403 1.193352 274.1525 13.27925 28.84136
#>  7      7 0.3007540 28.03180 270.60587 1.738948 583.0278 33.82573 50.33965
#>  8      8 1.2243960 29.22275 325.24820 1.795361 612.8485 40.65603 59.42775
#>  9      9 0.6697522 37.40394 341.08804 1.630735 525.2762 42.63600 66.04469
#> 10     10 2.5300553 24.72049 270.63366 1.578528 497.0904 33.82921 53.60120
#> 11     11 0.6090013 19.55393 132.11878 1.370547 381.4725 16.51485 29.60541
#> 12     12 1.5715652 23.26825 188.59710 1.392899 394.2512 23.57464 41.55828
#> 13     13 3.1519712 31.41656 440.07865 1.854158 643.7614 55.00983 79.05473
#> 14     14 1.5071344 14.06758  98.63679 1.280308 328.4862 12.32960 23.97794
#> 15     15 0.2816108 39.30136 258.09147 1.388279 391.6190 32.26143 57.06404
#> 16     16 1.1524354 22.25631 167.30735 1.385311 389.9253 20.91342 37.07290
#> 17     17 3.8987574 20.98113 308.28220 1.806138 618.5265 38.53528 56.14573
#> 18     18 0.4334551 34.97931 191.45794 1.255076 313.1439 23.93224 47.86029
#> 19     19 0.9316964 30.32649 220.95429 1.388558 391.7779 27.61929 48.84299
#> 20     20 1.1039195 19.12133 106.20801 1.183866 267.9220 13.27600 29.27030
#> # ... with 1 more variables: Cminss <dbl>